+++ app/processing/components/get-adjustment.js
@@ -1,0 +1,11 @@
+const { convertToPounds } = require('../../utility')
+
+const getAdjustment = (previousValue, newValue, adjustmentValue) => {
+  return {
+    currentValue: convertToPounds(previousValue),
+    newValue: convertToPounds(newValue),
+    adjustmentValue: convertToPounds(adjustmentValue)
+  }
+}
+
+module.exports = getAdjustment
