+++ app/processing/components/index.js
@@ -3,11 +3,15 @@ const getDetails = require('./get-details')
 const getDetailedFunding = require('./get-detailed-funding')
 const getScheme = require('./get-scheme')
 const getDetailedPayments = require('./get-detailed-payments')
+const getSchedule = require('./get-schedule')
+const getAdjustment = require('./get-adjustment')
 
 module.exports = {
   getAddress,
   getDetailedFunding,
   getDetails,
   getDetailedPayments,
-  getScheme
+  getScheme,
+  getSchedule,
+  getAdjustment
 }
